backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
sqlalchemy==2.0.44
flask==3.1.2
flask-cors==6.0.1
sqlalchemy==2.0.44
flask==3.1.2
flask-cors==6.0.1
sqlalchemy==2.0.44
flask==3.1.2
flask-cors==6.0.1
sqlalchemy==2.0.44
psycopg[binary]==3.3.1
paho-mqtt==2.1.0
python-dotenv==1.1.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pydantic==2.5.2
supabase==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
alembic==1.17.2